Checked in Olaf's 1.5.2 sources.  Here are the changes:
v 1.5.2 (17. december 2003):

- modified netclient for not to drop received data: use of syspollfn
  instead of clock to poll for incoming data, circular recv buffer

v 1.5 (18. october 2003):
- added some usefull features to arraycopy (i.e. copying just parts of
  an array and copying to specified position in destination array)
- new object: nchange
- IRIX 6.5 port (for GCC 3.3)
- OS X binary (Jaguar 10.2.6)

v 1.4 (22. may 2003):
- updated sources to compile with Pd0.37-test4
- new object: arraycopy

v 1.3 (12. april 2003):
- new objects: sync listfifo
- all setup routines renamed to maxlib_<object>_setup() to avoid name
  clashes, old names still work via class_addcreator()
- some improvements for the help files


git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@1394 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/maxlib/src/sync.c
@@ -27,7 +27,7 @@
 #include <stdio.h>
 #include <math.h>
 
-#define SYNC_MAX_SIZE (sizeof(unsigned int) * 8)
+#define SYNC_MAX_SIZE (int)(sizeof(unsigned int) * 8)
 
 static char *version = "sync v0.1, written by Olaf Matthes <olaf.matthes@gmx.de>\n"
                        "           based on sync from jMax";
